Refactor error handling
Currently, error handling information is collapsed down to the point of not
allowing the user to investigate or mitigate the issue.

Use thiserror to define a set of possible errors that can be produced when
generating an SBOM. (This is somewhat limited by the fact that cargo::core makes
use of anyhow, which prevents us from ding nice things with #[from])

Remove the code from main that does the collapsing and allow anyhow to present
the error information to the user. This is not necessarily what we want as a
final approach, but it is a step in the right direction.

Signed-off-by: Amelia Keibler <amelia.keibler@gmail.com>