fix: lookup the top level dependencies from the resolved graph
Instead of using the declared dependency, this change now uses to the
resolved dependency graph to look up the top level dependencies.

The difference is, that a declared dependency may be overridden. And so
the source no longer matches, which results in the dependency missing
from the final SBOM.

Using the resolved dependency graph, overriding dependencies is now
covered as well.

This also aligns with the "all dependencies" approach, which also uses
the resolved dependencies.

Closes: #364
Signed-off-by: Jens Reimann <jreimann@redhat.com>