Convert to `cargo metadata` as a backend (#496)
This PR switches to the API-stable cargo metadata as opposed to the previous direct dependency on unstable internal APIs of Cargo

It is mostly bug-compatible, with the actual fixes to come in subsequent PRs.

# Benefits

- Far smaller binary size
- Shorter compile times
- Reliance only on stable APIs will require far less maintenance going forward
- Download crates in parallel #489 (we got it for free by switching to cargo metadata)

# Fixes
One of the tests was creating an invalid dependency that was getting ignored by Cargo, but still included in the SBOM. This is now fixed, and I've updated the test to create a proper dependency.

# Behavior changes
The previous code was inconsistent on what kinds of dependencies it includes. It previously included normal only (not build!) in top-level only mode, and all kinds (even dev!) in all deps mode. Now all kinds of dependencies are included always - it's still wrong, but at least it's consistent. Proper filtering by mode and package is implemented in a subsequent PR: #498
Regressions

Reading configuration from Cargo.toml is not fully wired up, although its tests pass. I've skipped it for now and left a comment because this behavior seems to be a bad idea in the first place: #495