Expand SpdxExpression with bom-ref
This changes the `SpdxExpression` to a struct with fields `bom_ref` &
`expression` to support spec version 1.5.

There is now a spec type `Expression` instead of a `String` to serialize
a `SpdxExpression`. Unfortunately the JSON deserialization code for `LicenseChoice` does not check a few of the invalid license related
samples correctly. It cannot detect the issue in

```json
{
  "expression": "EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0",
  "license": {
    "name": "Apache License 2.0"
  }
}
```

where both `expression` & `license` are given.

* update license sample files in test to check bom_ref

Signed-off-by: Sebastian Ziebell <sebastian.ziebell@ferrous-systems.com>