Refactor & fix handling of external reference URI
This refactors the logic to convert & validate the `url` field of the
`ExternalReference`. Instead of converting the type during serialization
the URL or BOM-Link is read as a String first, similar to how all other
URLs are handled. Only during validation the content is checked
specifically, then it's determined if it's a URL or a BOM-Link.
BOM-Links are only supported in spec version 1.5, all associated tests
are updated to reflect this.

* refactor `url` to be of type `String` in `ExternalReference`
* adjust serialization code, no error is raised, validation is done
  separately
* fix validation of `Uri` type (that can be `Url` or `BomLink`) checking
  the provided `SpecVersion`
* add test to check validation of an `ExternalReference`
* remove a lot of code

Signed-off-by: Sebastian Ziebell <sebastian.ziebell@ferrous-systems.com>