Remove Bug Bounty program
We received almost entirely AI slop reports that are irrelevant to our tool. It's a library and most reporters didn't even bother to read the rules or even look at what the intended purpose of the tool is/was.

This caused a lot of extra work which is why we decided to abandon the program. Thanks AI.

Signed-off-by: Lars Francke <git@lars-francke.de>